requirements.txt
streamlit
pandas
openpyxl
PyYAML
langchain
sentence-transformers
faiss-cpu
transformers
torch